requirements.txt
Django>=1.8,<1.9
psycopg2
geojson
python-social-auth>=0.2.7
pykml
raven

django-ipware
django-reversion>=1.8.5
djorm-ext-pgarray==0.10
pyyaml

